Switched from CGE::Log to new tracing.
git-svn-id: svn://svn.code.sf.net/p/captainsengine/code/trunk@455 dc34a4ac-3e42-4537-918b-2dea8b01751e

+++ Adventure/AppBuilder/AppBuilder/ScriptFunc.cpp
@@ -1065,7 +1065,7 @@ int ScriptFunctions::setParticles(ExecutionContext& ctx, unsigned numArgs){
   float anglespeed = (float)(rotation/180.0f*M_PI);
   Engine::instance()->getParticleEngine()->setRotation(anglespeed);
   Engine::instance()->getParticleEngine()->setSpeedVariation(variation/100.0f);
-  Engine::instance()->getParticleEngine()->setParticleObject(object, angle-M_PI/2.0f);
+  Engine::instance()->getParticleEngine()->setParticleObject(object, float(angle-M_PI/2.0f));
   return 0;
 }
